+++ dmd2/mars.h
@@ -318,7 +318,7 @@ struct Compiler
     const char *vendor;     // Compiler backend name
 };
 
-typedef unsigned structalign_t;
+typedef signed long structalign_t;
 #define STRUCTALIGN_DEFAULT ~0  // magic value means "match whatever the underlying C compiler does"
 // other values are all powers of 2
